LayoutTests/imported/w3c/web-platform-tests/trusted-types/trusted-types-source-file-path-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'require-trusted-types-for'.

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'trusted-types'.


FAIL same-document script assert_throws_js: function "_ => {
    document.getElementById("to-be-modified").innerHTML = "'test'";
  }" did not throw
FAIL same-origin script assert_throws_js: function "() => setInnerHtml(toBeModified, "'test'")" did not throw
FAIL cross-origin script assert_throws_js: function "() => setInnerHtml(toBeModified, "'test'")" did not throw
